file deletion builds on windows
